+++ nosqldb/nson_serializer.go
@@ -1006,7 +1006,7 @@ func (req *WriteMultipleRequest) serialize(w proto.Writer, serialVersion int16) 
 		} else {
 			if strings.EqualFold(tableName, operation.tableName()) == false {
 				isSingleTable = false
-				break;
+				break
 			}
 		}
 	}
@@ -1043,11 +1043,11 @@ func (req *WriteMultipleRequest) serialize(w proto.Writer, serialVersion int16) 
 		if operation.DeleteRequest != nil {
 			operation.DeleteRequest.abortOnFail = operation.AbortOnFail
 			subReq = operation.DeleteRequest
-			err = operation.DeleteRequest.serializeInternal(w, serialVersion, isSingleTable==false)
+			err = operation.DeleteRequest.serializeInternal(w, serialVersion, isSingleTable == false)
 		} else if operation.PutRequest != nil {
 			operation.PutRequest.abortOnFail = operation.AbortOnFail
 			subReq = operation.PutRequest
-			err = operation.PutRequest.serializeInternal(w, serialVersion, isSingleTable==false)
+			err = operation.PutRequest.serializeInternal(w, serialVersion, isSingleTable == false)
 		}
 
 		if err != nil {